storage: Rename indexeddb_thread to indexeddb in storage_traits
The extra suffix is not necessary, as the separation between the script side
and the backend side of the IndexedDB implementation is already clear from
imports (dom::indexeddb vs storage_traits::indexeddb).

This patch renames indexeddb_thread.rs to indexeddb.rs and updates all imports
accordingly. No functional changes are included.

Testing: Unit tests continue to pass.
